pyproject.toml
[project]
name = "yolo-earth-segmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "ultralytics>=8.3.167",
    "uvicorn>=0.35.0",
]
